added hal to make main cleaner
